Started writing install.py Also moved python scripts to new folder to clean up things a little.
