+++ pyscripts/s000_detect_hardware.py
