+++ pyscripts/s07_hostname.py
