+++ pyscripts/s12_aur.py
